blogs_project/requirements.txt
asgiref==3.8.1
Django==5.1.1
django-filter==24.3
djangorestframework==3.15.2
Markdown==3.7
psycopg==3.2.2
sqlparse==0.5.1
typing_extensions==4.12.2
